backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
# ML / YOLO tracking e Visão Computacional
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
supervision>=0.18.0
easyocr>=1.7.0
deep-sort-realtime>=1.3.0
imageio-ffmpeg>=0.6.0

annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.133.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.14.0
fastar==0.8.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.0
psycopg2-binary==2.9.10
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.22
passlib[bcrypt]==1.7.4
PyYAML==6.0.3
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.53.0
setuptools==69.5.1
shellingham==1.5.4
SQLAlchemy==2.0.48
sqlmodel==0.0.37
starlette==0.52.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
wheel==0.46.3